+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Fixed unit tests.
 - SqlAGReplica
   - Fix unit test FailedRemoveAvailabilityGroupReplica
+- SqlAgentOperator
+  - Unit test for changing e-mail address on an existing operator.
 
 ### Changed